+++ src/wav_file.h
@@ -33,7 +33,7 @@ WAV_FILE* wav_open(const char* filename);
 void wav_read(WAV_FILE* wav_file, int16_t* buffer, int num_samples);
 
 WAV_FILE* wav_open_write(const char* filename, WAV_HEADER header);
-void wav_write(WAV_FILE* wav_file, int16_t** channel_buffers, int chunk_size);
+void wav_write(WAV_FILE* wav_file, double** channel_buffers, int chunk_size);
 
 void wav_close(WAV_FILE* wav_file);
